Moving some more oc commands to staging
