+++ hack/import-restrictions.json
@@ -412,14 +412,12 @@
       "github.com/openshift/origin/pkg/oc/cli/admin/groups/examples"
     ],
     "allowedImportPackageRoots": [
-      "vendor/github.com/containers/image",
       "vendor/github.com/docker/go-units",
       "vendor/github.com/fsnotify/fsnotify",
       "vendor/github.com/fsouza/go-dockerclient",
       "vendor/k8s.io/klog",
       "vendor/github.com/gonum/graph",
       "vendor/github.com/MakeNowJust/heredoc",
-      "vendor/github.com/opencontainers/go-digest",
       "vendor/github.com/openshift/api",
       "vendor/github.com/openshift/client-go",
       "vendor/github.com/openshift/library-go",
@@ -443,7 +441,6 @@
       "vendor/github.com/openshift/oc"
     ],
     "allowedImportPackages": [
-      "github.com/openshift/origin/pkg/api",
       "github.com/openshift/origin/pkg/build/apis/build",
       "github.com/openshift/origin/pkg/build/apis/build/install",
       "github.com/openshift/origin/pkg/cmd/templates",
@@ -452,7 +449,6 @@
       "github.com/openshift/origin/pkg/version",
 
       "vendor/k8s.io/kubernetes/pkg/api/legacyscheme",
-      "vendor/k8s.io/kubernetes/pkg/credentialprovider",
 
       "vendor/github.com/openshift/library-go/pkg/git"
     ]
