+++ pkg/oc/cli/cli.go
@@ -36,7 +36,10 @@ import (
 	"github.com/openshift/oc/pkg/cli/observe"
 	"github.com/openshift/oc/pkg/cli/options"
 	"github.com/openshift/oc/pkg/cli/process"
+	"github.com/openshift/oc/pkg/cli/project"
+	"github.com/openshift/oc/pkg/cli/projects"
 	"github.com/openshift/oc/pkg/cli/recycle"
+	"github.com/openshift/oc/pkg/cli/requestproject"
 	"github.com/openshift/oc/pkg/cli/rollback"
 	"github.com/openshift/oc/pkg/cli/rsh"
 	"github.com/openshift/oc/pkg/cli/secrets"
@@ -52,10 +55,7 @@ import (
 	"github.com/openshift/origin/pkg/oc/cli/newapp"
 	"github.com/openshift/origin/pkg/oc/cli/newbuild"
 	"github.com/openshift/origin/pkg/oc/cli/policy"
-	"github.com/openshift/origin/pkg/oc/cli/project"
-	"github.com/openshift/origin/pkg/oc/cli/projects"
 	"github.com/openshift/origin/pkg/oc/cli/registry"
-	"github.com/openshift/origin/pkg/oc/cli/requestproject"
 	"github.com/openshift/origin/pkg/oc/cli/rollout"
 	"github.com/openshift/origin/pkg/oc/cli/rsync"
 	"github.com/openshift/origin/pkg/oc/cli/set"
