+++ staging/src/github.com/openshift/oc/pkg/cli/admin/migrate/images/imagerefs_test.go
