+++ staging/src/github.com/openshift/oc/pkg/cli/admin/verifyimagesignature/manifest.go
