+++ staging/src/github.com/openshift/oc/pkg/cli/project/project.go
@@ -18,11 +18,11 @@ import (
 	kcmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 	"k8s.io/kubernetes/pkg/kubectl/util/templates"
 
+	"github.com/openshift/api/annotations"
 	projectv1 "github.com/openshift/api/project/v1"
 	projectv1client "github.com/openshift/client-go/project/clientset/versioned/typed/project/v1"
 	cliconfig "github.com/openshift/oc/pkg/helpers/kubeconfig"
 	clientcfg "github.com/openshift/oc/pkg/helpers/originkubeconfignames"
-	oapi "github.com/openshift/origin/pkg/api"
 )
 
 type ProjectOptions struct {
@@ -374,7 +374,7 @@ func convertNamespaceList(namespaceList *corev1.NamespaceList) *projectv1.Projec
 }
 
 func DisplayNameForProject(project *projectv1.Project) string {
-	displayName := project.Annotations[oapi.OpenShiftDisplayName]
+	displayName := project.Annotations[annotations.OpenShiftDisplayName]
 	if len(displayName) == 0 {
 		displayName = project.Annotations["displayName"]
 	}
