+++ test/integration/unprivileged_newproject_test.go
@@ -16,8 +16,8 @@ import (
 	authorizationv1client "github.com/openshift/client-go/authorization/clientset/versioned"
 	projectv1client "github.com/openshift/client-go/project/clientset/versioned/typed/project/v1"
 	templatev1client "github.com/openshift/client-go/template/clientset/versioned/typed/template/v1"
+	"github.com/openshift/oc/pkg/cli/requestproject"
 	"github.com/openshift/oc/pkg/helpers/tokencmd"
-	"github.com/openshift/origin/pkg/oc/cli/requestproject"
 	projectapi "github.com/openshift/origin/pkg/project/apis/project"
 	testutil "github.com/openshift/origin/test/util"
 	testserver "github.com/openshift/origin/test/util/server"
